third_party/WebKit/LayoutTests/battery-status/no-gc-with-eventlisteners-expected.txt

Test no garbage collection of battery manager object when listeners are attached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS batteryInfo is defined.
PASS lastSetMockBatteryInfo is defined.
PASS batteryInfo.charging is lastSetMockBatteryInfo.charging
PASS batteryInfo.chargingTime is lastSetMockBatteryInfo.chargingTime
PASS batteryInfo.dischargingTime is lastSetMockBatteryInfo.dischargingTime
PASS batteryInfo.level is lastSetMockBatteryInfo.level
PASS successfullyParsed is true

TEST COMPLETE
